refactor(chat): migrer les pages et supprimer l'ancien répertoire chat/
Met à jour les imports dans les pages chat vers les nouveaux emplacements
(organisms/chat/chat-panel). Supprime les anciens composants et tests
dans components/chat/ désormais remplacés par la hiérarchie atomic design.
